merges openable container and storage container in one script
Those two scripts were doing very similar things, making the container interactable.
