+++ Assets/Content/Furniture/Machines/Kitchen/Microwave/Microwave.cs
@@ -20,8 +20,6 @@ public class Microwave : InteractionTargetNetworkBehaviour
     // Prefab for the ashes, when an item is burned
     // TODO: implement the assetdata stuffs
     public GameObject DestroyedItemPrefab;
-    // Place for the container to place items that is placed in the object
-    public AttachedContainer AttachedContainer;
 
     private AudioSource audioSource;
     // Sound that plays when its turned on
@@ -38,13 +36,12 @@ public class Microwave : InteractionTargetNetworkBehaviour
     // we might have isOn for electricity stuff
     private bool isOn;
     // actual container
-    private OpenableContainer openableContainer;
+    public ContainerDescriptor containerDescriptor;
 
     private void Start()
     {
-        Assert.IsNotNull(AttachedContainer);
 
-        openableContainer = GetComponent<OpenableContainer>();
+        Assert.IsNotNull(containerDescriptor);
         audioSource = GetComponent<AudioSource>();
 
         emissionMaterial = GetComponent<Renderer>().materials[1];
@@ -70,10 +67,10 @@ public class Microwave : InteractionTargetNetworkBehaviour
         }
 
         // Can't be turned on if the door is open, we might add a hacking thing to bypass this later
-        if (openableContainer != null && openableContainer.IsOpen())
-        {
-            return false;
-        }
+        //if (openableContainer != null && openableContainer.IsOpen())
+        //{
+        //    return false;
+        //}
 
         return !isOn;
     }
@@ -90,10 +87,10 @@ public class Microwave : InteractionTargetNetworkBehaviour
     private void SetActivated(bool activated)
     {
         isOn = activated;
-        if (openableContainer != null)
-        {
-            openableContainer.enabled = !activated;
-        }
+       // if (openableContainer != null)
+      //  {
+      //      openableContainer.enabled = !activated;
+      //  }
     }
 
     // Start a cycle
@@ -110,7 +107,7 @@ public class Microwave : InteractionTargetNetworkBehaviour
 
     private void CookItems()
     {
-        var items = AttachedContainer.Container.Items.ToArray();
+        var items = containerDescriptor.attachedContainer.Container.Items.ToArray();
 
         // tries to get a microweavable in each item that is in the container
         foreach (Item item in items)
