+++ Assets/Content/Furniture/Machines/Supply/DisposalBin/DisposalBin.prefab
@@ -16,11 +16,12 @@ GameObject:
   - component: {fileID: 4265221962216021434}
   - component: {fileID: 6707010917967640726}
   - component: {fileID: 7091454980463605699}
-  - component: {fileID: 8480585871254921602}
-  - component: {fileID: 7571045420630100205}
-  - component: {fileID: 1617368548363267037}
   - component: {fileID: 4859281451868802076}
-  - component: {fileID: 1824158099739291962}
+  - component: {fileID: 7709656812663945726}
+  - component: {fileID: 2460959987605775005}
+  - component: {fileID: 8139242482030902464}
+  - component: {fileID: 7545054435155777865}
+  - component: {fileID: 3151284307672784910}
   m_Layer: 0
   m_Name: DisposalBin
   m_TagString: Untagged
@@ -275,60 +276,6 @@ AudioSource:
     m_PreInfinity: 2
     m_PostInfinity: 2
     m_RotationOrder: 4
---- !u!114 &8480585871254921602
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6905024007199924851}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 25696236470760b41b375e9b6939444c, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  attachedContainer: {fileID: 7571045420630100205}
-  attachedContainerGenerator: {fileID: 1617368548363267037}
-  containerSync: {fileID: 1824158099739291962}
-  openableContainer: {fileID: 0}
-  storageContainer: {fileID: 4859281451868802076}
-  visibleContainer: {fileID: 0}
-  containerName: main garbage
-  onlyStoreWhenOpen: 0
-  size: {x: 100, y: 100}
-  hideItems: 1
-  attachItems: 1
-  startFilter: {fileID: 0}
-  initialized: 1
-  maxDistance: 5
-  containerType: 1
---- !u!114 &7571045420630100205
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6905024007199924851}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 72bfc9c5e18140bc9dbf232d3a7c2ff6, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  AttachmentOffset: {x: 0, y: 0, z: 0}
-  containerDescriptor: {fileID: 8480585871254921602}
---- !u!114 &1617368548363267037
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6905024007199924851}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: fa364422073946a1badbfa66a349361d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  containerDescriptor: {fileID: 8480585871254921602}
 --- !u!114 &4859281451868802076
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -343,9 +290,83 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   syncMode: 0
   syncInterval: 0.1
-  containerDescriptor: {fileID: 8480585871254921602}
-  viewContainerIcon: {fileID: 0}
---- !u!114 &1824158099739291962
+  containerDescriptor: {fileID: 0}
+--- !u!114 &7709656812663945726
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6905024007199924851}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 25696236470760b41b375e9b6939444c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  attachedContainer: {fileID: 2460959987605775005}
+  attachedContainerGenerator: {fileID: 8139242482030902464}
+  containerSync: {fileID: 3151284307672784910}
+  containerInteractive: {fileID: 7545054435155777865}
+  visibleContainer: {fileID: 0}
+  openIcon: {fileID: 0}
+  takeIcon: {fileID: 0}
+  storeIcon: {fileID: 0}
+  viewIcon: {fileID: 0}
+  attachmentOffset: {x: 0, y: 0, z: 0}
+  containerName: disposal
+  onlyStoreWhenOpen: 0
+  size: {x: 10, y: 10}
+  hideItems: 1
+  attachItems: 1
+  startFilter: {fileID: 0}
+  initialized: 1
+  maxDistance: 5
+  containerType: 1
+  isOpenable: 0
+  isInteractive: 1
+--- !u!114 &2460959987605775005
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6905024007199924851}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 72bfc9c5e18140bc9dbf232d3a7c2ff6, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  containerDescriptor: {fileID: 7709656812663945726}
+--- !u!114 &8139242482030902464
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6905024007199924851}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: fa364422073946a1badbfa66a349361d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  containerDescriptor: {fileID: 7709656812663945726}
+--- !u!114 &7545054435155777865
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6905024007199924851}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3c75d17834848bb4f83ee6e730a65519, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  syncMode: 0
+  syncInterval: 0.1
+  OpenIcon: {fileID: 0}
+  containerDescriptor: {fileID: 0}
+--- !u!114 &3151284307672784910
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
