+++ Assets/Content/Furniture/Storage/ContainerInteractive.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3c75d17834848bb4f83ee6e730a65519
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
