+++ Assets/Content/Items/Functional/Containers/Object/Toolboxes/RedToolbox.prefab
@@ -231,11 +231,13 @@ GameObject:
   - component: {fileID: 2884915648782529247}
   - component: {fileID: -8822467730395627730}
   - component: {fileID: 5206759755089209485}
-  - component: {fileID: -5974139538508048618}
   - component: {fileID: -396234361631906964}
-  - component: {fileID: 532509687929333354}
-  - component: {fileID: 1093985765880061407}
-  - component: {fileID: 2103983410208154681}
+  - component: {fileID: 8984766001432014251}
+  - component: {fileID: 4469562446528477193}
+  - component: {fileID: 4064354730430873157}
+  - component: {fileID: 7681387737837336857}
+  - component: {fileID: 2619326083095352681}
+  - component: {fileID: 8774879036025570029}
   m_Layer: 16
   m_Name: RedToolbox
   m_TagString: Untagged
@@ -364,26 +366,6 @@ Animator:
   m_HasTransformHierarchy: 1
   m_AllowConstantClipSamplingOptimization: 1
   m_KeepAnimatorControllerStateOnDisable: 0
---- !u!114 &-5974139538508048618
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5533951594911835035}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3c75d17834848bb4f83ee6e730a65519, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  syncMode: 0
-  syncInterval: 0.1
-  OpenIcon: {fileID: 21300000, guid: ce4cb1c9f32561c4db2ae13a08e28987, type: 3}
-  StorageType: 0
-  Pile: 0
-  OnlyStoreWhenOpen: 1
-  MaxDistance: 1.5
-  viewContainerIcon: {fileID: 21300000, guid: 0f4328c69a4a89541a87f2482052e497, type: 3}
 --- !u!114 &-396234361631906964
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -402,7 +384,40 @@ MonoBehaviour:
   localPositionSensitivity: 0.01
   localRotationSensitivity: 0.01
   localScaleSensitivity: 0.01
---- !u!114 &532509687929333354
+--- !u!114 &8984766001432014251
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5533951594911835035}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 25696236470760b41b375e9b6939444c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  attachedContainer: {fileID: 4469562446528477193}
+  attachedContainerGenerator: {fileID: 4064354730430873157}
+  containerSync: {fileID: 8774879036025570029}
+  containerInteractive: {fileID: 7681387737837336857}
+  visibleContainer: {fileID: 2619326083095352681}
+  openIcon: {fileID: 0}
+  takeIcon: {fileID: 0}
+  storeIcon: {fileID: 0}
+  viewIcon: {fileID: 0}
+  attachmentOffset: {x: 0, y: 0, z: 0}
+  containerName: container
+  onlyStoreWhenOpen: 1
+  size: {x: 3, y: 3}
+  hideItems: 1
+  attachItems: 1
+  startFilter: {fileID: 0}
+  initialized: 1
+  maxDistance: 5
+  containerType: 0
+  isOpenable: 1
+  isInteractive: 1
+--- !u!114 &4469562446528477193
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -414,10 +429,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 72bfc9c5e18140bc9dbf232d3a7c2ff6, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  HideItems: 1
-  AttachItems: 1
-  AttachmentOffset: {x: 0, y: 0, z: 0}
---- !u!114 &1093985765880061407
+  containerDescriptor: {fileID: 8984766001432014251}
+--- !u!114 &4064354730430873157
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -429,10 +442,37 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: fa364422073946a1badbfa66a349361d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  AttachedContainer: {fileID: 532509687929333354}
-  Size: {x: 4, y: 3}
-  StartFilter: {fileID: 11400000, guid: 89eaffbef3faadd4ca026f279e092c8b, type: 2}
---- !u!114 &2103983410208154681
+  containerDescriptor: {fileID: 8984766001432014251}
+--- !u!114 &7681387737837336857
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5533951594911835035}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3c75d17834848bb4f83ee6e730a65519, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  syncMode: 0
+  syncInterval: 0.1
+  OpenIcon: {fileID: 0}
+  containerDescriptor: {fileID: 8984766001432014251}
+--- !u!114 &2619326083095352681
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5533951594911835035}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e3ca34599d24491183415a2fdc4fc033, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  containerDescriptor: {fileID: 8984766001432014251}
+--- !u!114 &8774879036025570029
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
