--- Assets/Engine/Inventory/ContainerDescriptorEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5bdf112ab04037a438bbe3f77047e2e2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
