+++ Assets/Engine/Inventory/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 41856b99dd32d8840a3e598ab2fb51db
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
